disk-menu - debug version -first edit
Disk-menu
I've a small M.2 drive that is verry fast, where I've 2 partition of aprox. 10Gb each on it. The left 100Gb is my /srv mount.
One should be for swap mount and the other for tmp/
I would like to have a menu to select it so the settings are prepared to be ready for next reboot.

First step:
Attempt to create a menu with available drives for swapon mounts in order to selected the wanted drive and replace the default swap.

Notes:
$buid : the unmounted device is my swap - Could I use this to list available mounts UUID but unmounted drives are not listed (my already mounted swap in my test)

Next steps:
 - I should add create resume for the specified device
 - I should add the swapon commands
 - I need to find a way to show the UUID in the menu choice instead of the device only